先輩にGUI追加してもらいました
敵増やした状態のままプッシュしましたすんません

+++ SourceCode/Coin.h
@@ -6,31 +6,31 @@
 class Coin
 {
 private:
-	static int CoinNum;	//現在の合計枚数
-	static int moveCoin;//動く枚数
-	static int gotCoin;	//総獲得コイン数
+	static long int CoinNum;	//現在の合計枚数
+	static long int moveCoin;//動く枚数
+	static long int gotCoin;	//総獲得コイン数
 
 public:
-	static int GetCoinNum() { return CoinNum; }				//ゲッター
-	static void SetCoinNum(int coin) { CoinNum = coin; }	//セッター（多分使わない）
+	static long int GetCoinNum() { return CoinNum; }				//ゲッター
+	static void SetCoinNum(long int coin) { CoinNum = coin; }	//セッター（多分使わない）
 
-	static void AddCoinNum(int add) { CoinNum += add;  music::play(coinGet);}		//加算
-	static void DegCoinNum(int deg) { CoinNum -= deg; }		//減算
+	static void AddCoinNum(long int add) { CoinNum += add;  music::play(coinGet);}		//加算
+	static void DegCoinNum(long int deg) { CoinNum -= deg; }		//減算
 
-	static int GetRatioCoin(float ratio);	//割合で取りだす
+	static long int GetRatioCoin(float ratio);	//割合で取りだす
 											//攻撃で使用した場合、しっかり減算してね
 
-	static int RewardCoin(float ratio) { return moveCoin * ratio; }	//返ってくるコイン
-	static int LightAttackReward() { return RewardCoin(lightRatio); }
-	static int HeavyAttackReward() { return RewardCoin(heavyRatio); }
+	static long int RewardCoin(float ratio) { return moveCoin * ratio; }	//返ってくるコイン
+	static long int LightAttackReward() { return RewardCoin(lightRatio); }
+	static long int HeavyAttackReward() { return RewardCoin(heavyRatio); }
 	static float lightRatio;
 	static float heavyRatio;
 
 	//ダメージ計算
-	static int calcDamage(int baseDamage, int coin) { return baseDamage * coin; }
+	static long int calcDamage(long int baseDamage, long int coin) { return baseDamage * coin; }
 
 	static void AddGotCoin(int add) { gotCoin += add; }	//獲得した枚数を入れる
-	static int GetGotCoin() { return gotCoin; }			//獲得した枚数のゲッター
+	static long int GetGotCoin() { return gotCoin; }			//獲得した枚数のゲッター
 	static void ResetGotCoin() { gotCoin = 0; }			//獲得したコインを０に戻す
 
 };
